requirements.txt
speechbrain
webvtt-py
youtube-dl
scikit-learn
inaSpeechSegmenter 
librosa
